Convert LWRPs to poise HWRPs
First pass at converting to poise-style HWRPs. The firewall resource has many firewall_rule subresources now, and the rule resources are simply accumulated until the end of the chef run, at which point the firewall resource triggers a full rebuild of the rules, looks for any changes compared to the current system, and behaves accordingly.