Expose default table and chains for iptables
This is a backwards compatible new feature to override the default ruleset, or add additional rules. I've also placed an example in the README.md for how to add rules using the `raw` parameter of `firewall_rule`. Eventually, I'd like to clean up this cookbook to only expose the *filter table by default, and ask users to rely on raw rules for anything beyond that. Supporting nat, security, mangle, and other tables is probably not going to translate cleanly to firewalld and ufw and the Windows firewall.

Fixes #101.