Various fixes for firewalld resources.
With this commit, values for properties of some firewalld resources
have to be specified in a more human readable way.

Examples:
- `'22/tcp'` instead of `[['22', 'tcp']]`
- `'port=8080:proto=tcp:toport=80:toaddr=192.0.2.1'` instead of
  `[['8080', 'tcp', '80', '192.0.2.1']]`

This is also consistent with how these values have to be presented to
`firewall-cmd`.  Some bugs and inconsistencies have been fixed along
the way.  Tests and docs have also been adjusted.

Do not use begin/rescue blocks when adding firewalld-objects, as that
resulted in errors being logged by firewalld.  Rather, the existence
of objects is now queried with if/else-blocks via the dbus-interface
of firewalld.

Signed-off-by: Matthias Pausch <m.pausch@gsi.de>